backend/requirements.txt
-i https://pypi.org/simple
annotated-types==0.7.0; python_version >= '3.8'
argcomplete==3.6.3; python_version >= '3.8'
blinker==1.9.0; python_version >= '3.9'
boto3==1.42.49; python_version >= '3.9'
botocore==1.42.49; python_version >= '3.9'
certifi==2026.1.4; python_version >= '3.7'
cfn-flip==1.3.0
charset-normalizer==3.4.4; python_version >= '3.7'
click==8.3.1; python_version >= '3.10'
durationpy==0.10
flask==3.1.2; python_version >= '3.9'
hjson==3.1.0
idna==3.11; python_version >= '3.8'
itsdangerous==2.2.0; python_version >= '3.8'
jinja2==3.1.6; python_version >= '3.7'
jmespath==1.1.0; python_version >= '3.9'
markupsafe==3.0.3; python_version >= '3.9'
packaging==26.0; python_version >= '3.8'
pip==26.0.1; python_version >= '3.9'
placebo==0.9.0
pydantic==2.12.5; python_version >= '3.9'
pydantic-core==2.41.5; python_version >= '3.9'
python-dateutil==2.9.0.post0; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2'
python-dotenv==1.2.1; python_version >= '3.9'
python-slugify==8.0.4; python_version >= '3.7'
python-ulid==3.1.0; python_version >= '3.9'
pyyaml==6.0.3; python_version >= '3.8'
requests==2.32.5; python_version >= '3.9'
s3transfer==0.16.0; python_version >= '3.9'
setuptools==79.0.1; python_version >= '3.9'
six==1.17.0; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2'
text-unidecode==1.3
toml==0.10.2; python_version >= '2.6' and python_version not in '3.0, 3.1, 3.2'
tqdm==4.67.3; python_version >= '3.7'
troposphere==4.10.0; python_version >= '3.8'
typing-extensions==4.15.0; python_version >= '3.9'
typing-inspection==0.4.2; python_version >= '3.9'
urllib3==2.6.3; python_version >= '3.9'
werkzeug==3.1.5; python_version >= '3.9'
wheel==0.46.3; python_version >= '3.9'
zappa==0.61.4; python_version >= '3.9'
